src/matsci_llm_causality/extraction/pdf.py
"""
PDF processing and text extraction module.
"""

from pathlib import Path
from typing import List, Optional
import fitz  # PyMuPDF

class PDFProcessor:
    """Handles PDF document processing and text extraction."""
    
    def extract_text(self, pdf_path: str | Path) -> str:
        """
        Extract text content from a PDF file.
        
        Args:
            pdf_path: Path to the PDF file
            
        Returns:
            Extracted text content
        """
        doc = fitz.open(pdf_path)
        text = []
        for page in doc:
            text.append(page.get_text())
        return "\n".join(text)
    
    def extract_sections(self, pdf_path: str | Path) -> dict[str, str]:
        """
        Extract text content by section from a PDF file.
        
        Args:
            pdf_path: Path to the PDF file
            
        Returns:
            Dictionary mapping section names to their text content
        """
        # TODO: Implement section detection and extraction
        return {"full_text": self.extract_text(pdf_path)}
