pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "matsci_llm_causality"
version = "0.1.0"
description = "Extract causal relationships from materials science literature using LLMs"
readme = "README.md"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["materials-science", "causal-discovery", "llm", "nlp"]
dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "PyMuPDF>=1.20.0",
    "spacy>=3.5.0",
    "networkx>=3.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.0.0",
    "pandas>=2.0.0",
    "numpy>=1.20.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=22.0",
    "isort>=5.0",
    "mypy>=1.0",
    "ruff>=0.0.1",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/matsci_llm_causality"
Documentation = "https://matsci-llm-causality.readthedocs.io/"
Repository = "https://github.com/yourusername/matsci_llm_causality.git"

[tool.hatch.build.targets.wheel]
packages = ["src/matsci_llm_causality"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=matsci_llm_causality"

[tool.black]
line-length = 88
target-version = ["py39", "py310"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C"]
